Plots and PTR Running but Not Converging Yet
All code is there for PTR and there are now plotting functions to show that the linearization and discretization is working (I seems) so issues #6 and #7 are closed. #8 just needs to be debugged. Potentially it is an issues with the PTR solver parameters but I would bet it is some error with how things are indexed or something.