PTR Working 🚀  (small improvements needed)
Gets a solution which is slightly suboptimal compared to CasADi but converges well given a bad initial guess. Right now you need to manually set the trust region weights but by making scaling of the problem variables happen correctly, the automatic trust region weight update should work making the algorithm more automatic. Inspirations from the SCPToolbox.jl implementation of PTR should be tried to make the optimality better. FOH is implemented and works but it needs to be seem if it is working as well as it should. Issues #8 , #7 , #13 completed.